+++ angular_test/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.0.2-alpha+1
+
 -   Use the new generic function syntax, stop using `package:func`.
 -   Support breaking changes and deprecations in angular 5.0.0-alpha+1.